requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.22.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.1
pytest-flask==1.1.0
python-dotenv==0.15.0
toml==0.10.2
Werkzeug==1.0.1
